refactor(macos): isolate clash config generation
